nlp_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

torch>=2.0.0
transformers==4.35.2
sentence-transformers==2.2.2
keybert==0.7.0

nltk==3.8.1
easyocr==1.7.1
opencv-python-headless==4.8.1.78
numpy<2.0.0
Pillow==10.1.0